requirements.txt
selenium~=4.32.0
pytest~=8.3.5
pyyaml~=6.0.2
allure-pytest
allure-python-commons~=2.14.2
pytest-rerunfailures
pytest-html
loguru~=0.7.2
pytest-xdist
pytest-order
pytest-dependency
python-dotenv
